requirements.txt
fpdf
requests
beautifulsoup4
flask
Flask-SQLAlchemy
Pillow
